@@ -1,19 +1,20 @@
 {
   "name": "path-scurry",
-  "version": "1.10.1",
+  "version": "1.11.1",
   "description": "walk paths fast and efficiently",
   "author": "Isaac Z. Schlueter <i@izs.me> (https://blog.izs.me)",
-  "main": "./dist/cjs/index.js",
-  "module": "./dist/mjs/index.js",
+  "main": "./dist/commonjs/index.js",
+  "type": "module",
   "exports": {
+    "./package.json": "./package.json",
     ".": {
       "import": {
-        "types": "./dist/mjs/index.d.ts",
-        "default": "./dist/mjs/index.js"
+        "types": "./dist/esm/index.d.ts",
+        "default": "./dist/esm/index.js"
       },
       "require": {
-        "types": "./dist/cjs/index.d.ts",
-        "default": "./dist/cjs/index.js"
+        "types": "./dist/commonjs/index.d.ts",
+        "default": "./dist/commonjs/index.js"
       }
     }
   },
@@ -25,18 +26,17 @@
     "preversion": "npm test",
     "postversion": "npm publish",
     "prepublishOnly": "git push origin --follow-tags",
-    "preprepare": "rm -rf dist",
-    "prepare": "tsc -p tsconfig.json && tsc -p tsconfig-esm.json",
-    "postprepare": "bash ./scripts/fixup.sh",
+    "prepare": "tshy",
     "pretest": "npm run prepare",
     "presnap": "npm run prepare",
-    "test": "c8 tap",
-    "snap": "c8 tap",
+    "test": "tap",
+    "snap": "tap",
     "format": "prettier --write . --loglevel warn",
     "typedoc": "typedoc --tsconfig tsconfig-esm.json ./src/*.ts",
     "bench": "bash ./scripts/bench.sh"
   },
   "prettier": {
+    "experimentalTernaries": true,
     "semi": false,
     "printWidth": 75,
     "tabWidth": 2,
@@ -47,31 +47,25 @@
     "arrowParens": "avoid",
     "endOfLine": "lf"
   },
-  "tap": {
-    "coverage": false,
-    "node-arg": [
-      "--no-warnings",
-      "--loader",
-      "ts-node/esm"
-    ],
-    "ts": false
-  },
   "devDependencies": {
     "@nodelib/fs.walk": "^1.2.8",
-    "@types/node": "^20.1.4",
-    "@types/tap": "^15.0.7",
+    "@types/node": "^20.12.11",
     "c8": "^7.12.0",
     "eslint-config-prettier": "^8.6.0",
     "mkdirp": "^3.0.0",
-    "prettier": "^2.8.3",
+    "prettier": "^3.2.5",
     "rimraf": "^5.0.1",
-    "tap": "^16.3.4",
-    "ts-node": "^10.9.1",
-    "typedoc": "^0.23.24",
-    "typescript": "^5.0.4"
+    "tap": "^18.7.2",
+    "ts-node": "^10.9.2",
+    "tshy": "^1.14.0",
+    "typedoc": "^0.25.12",
+    "typescript": "^5.4.3"
+  },
+  "tap": {
+    "typecheck": true
   },
   "engines": {
-    "node": ">=16 || 14 >=14.17"
+    "node": ">=16 || 14 >=14.18"
   },
   "funding": {
     "url": "https://github.com/sponsors/isaacs"
@@ -81,7 +75,15 @@
     "url": "git+https://github.com/isaacs/path-scurry"
   },
   "dependencies": {
-    "lru-cache": "^9.1.1 || ^10.0.0",
+    "lru-cache": "^10.2.0",
     "minipass": "^5.0.0 || ^6.0.2 || ^7.0.0"
-  }
+  },
+  "tshy": {
+    "selfLink": false,
+    "exports": {
+      "./package.json": "./package.json",
+      ".": "./src/index.ts"
+    }
+  },
+  "types": "./dist/commonjs/index.d.ts"
 }
