@@ -1,9 +1,13 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-    content: [],
-    theme: {
-        extend: {},
-    },
-    plugins: [require('flowbite/plugin')],
-    darkMode: 'class',
-};
+export default {
+  content: [
+    "./*.{html,js}",
+    "./node_modules/flowbite/**/*.js",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [require('flowbite/plugin')],
+  darkMode: 'class',
+}
+
